Move Placeholder classes to own files
